+++ semcore/date-picker/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 ### Added
 
 - `showError` property to `*Comparator.ValueDateRange` and `*Comparator.CompareDateRange` components.
+- Render children in `MonthDateRangeComparator.RangeCalendar`.
 
 ## [4.39.0] - 2024-04-29